backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
transformers==4.45.0
torch==2.6.0
psycopg2-binary==2.9.9
sentencepiece==0.2.1
